+++ cinema/main/admin.py
