+++ cinema/main/templates/main/showtime_list.html
