+++ cinema/main/views.py
